+++ Demo/KeyboardTextInput/DemoKeyboardView.swift
@@ -48,8 +48,10 @@ struct DemoKeyboardView: View {
             SystemKeyboard(
                 controller: controller,
                 autocompleteToolbar: .none,
-                buttonContent: { $1 },
-                buttonView: { $1 }
+                buttonContent: { $0.view },
+                buttonView: { $0.view },
+                emojiKeyboard: { _ in Color.red },
+                toolbar: { $0.view }
             )
         }.buttonStyle(.plain)
     }