+++ Tests/KeyboardKitTests/Mocks/MockDragGestureHandler.swift
@@ -6,14 +6,25 @@
 //
 
 import CoreGraphics
-import KeyboardKit
 import MockingKit
+@testable import KeyboardKit
 
 class MockDragGestureHandler: Mock, DragGestureHandler {
     
     lazy var handleDragGestureRef = MockReference(handleDragGesture)
     
     func handleDragGesture(from startLocation: CGPoint, to currentLocation: CGPoint) {
+        call(handleDragGestureRef, args: (startLocation, currentLocation))
+    }
+}
+
+class MockSpaceDragGestureHandler: Gestures.SpaceDragGestureHandler, Mockable {
+
+    var mock = Mock()
+    
+    lazy var handleDragGestureRef = MockReference(handleDragGesture)
+    
+    override func handleDragGesture(from startLocation: CGPoint, to currentLocation: CGPoint) {
         call(handleDragGestureRef, args: (startLocation, currentLocation))
     }
 }